fixed help and add directory check for remove dir function

+++ utils/file.py
@@ -49,9 +49,10 @@ def remove_dirs_except(path, dir_exception_list = {}):
 
     for dir in dir_list:
         target_path = os.path.join(path, dir)
-
-        if dir not in dir_exception_list:
-            shutil.rmtree(target_path)
+        
+        if os.path.isdir(target_path):
+            if dir not in dir_exception_list:
+                shutil.rmtree(target_path)
 
 
 def make_dir_if_not_exists(path):
